feat: [#73] 레이블 수정 구현
- 새 레이블 추가 / 기존 레이블 수정 View Controller의 중복 코드를 SuperClass로 두고 각각 SubClassing함
- 코드 재사용성 극대화를 위해 위와 같이 수정
